backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0

# LangChain ecosystem
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
langchain-google-genai==2.0.0
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langchain-classic==0.1.0

# Vector store
chromadb==0.5.0

# YouTube
youtube-transcript-api==1.0.0
google-api-python-client==2.130.0

# Utilities
httpx==0.27.0
